# types: intern AccAddress.String() to gut wasted expensive recomputations