@odeke-em 2021-02-25
It caches that result so that we’ll never have to even invoke .Empty().
Think of this change as basically running the code as it was before, but
this time under a cache to avoid recomputations.

On Wed, Feb 24, 2021 at 4:24 PM Alessio Treglia <notifications@github.com>
wrote:

> *@alessio* commented on this pull request.
> ------------------------------
>
> In types/address.go
> <https://github.com/cosmos/cosmos-sdk/pull/8694#discussion_r582401828>:
>
> >  // String implements the Stringer interface.
> -func (aa AccAddress) String() string {
> +func (aa AccAddress) String() (str string) {
> +	addMu.Lock()
> +	defer addMu.Unlock()
> +
> +	if str, ok := addrStrMemoize[string(aa)]; ok {
> +		return str
> +	}
> +
> +	// Otherwise cache it for later memoization.
> +	defer func() {
> +		addrStrMemoize[string(aa)] = str
> +	}()
> +
>
> Shouldn't the following block come always first:
>
> if aa.Empty() {
> 	return ""
> }
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/cosmos/cosmos-sdk/pull/8694#pullrequestreview-598053863>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABFL3V3HDPUREWRWO2L4WXLTAWKCVANCNFSM4YFPGATA>
> .
>