@odeke-em 2021-02-25
That can come later on, for now this suffices. The eviction mechanism would
be perhaps LRU based with a memory bound. Progressively we can get to the
final solution if need be :-) I plan on running nodes under live traffic
with continuous profiling so we’ll figure out what needs to be fixed.

On Thu, Feb 25, 2021 at 1:16 AM Marko <notifications@github.com> wrote:

> *@marbar3778* commented on this pull request.
> ------------------------------
>
> In types/address.go
> <https://github.com/cosmos/cosmos-sdk/pull/8694#discussion_r582666183>:
>
> > @@ -236,8 +237,28 @@ func (aa AccAddress) Bytes() []byte {
>  	return aa
>  }
>
> +// AccAddress.String() is expensive and if unoptimized dominantly showed up in profiles,
> +// yet has no mechanisms to trivially cache the result given that AccAddress is a []byte type.
> +// With the string interning below, we are able to achieve zero cost allocations for string->[]byte
> +// because the Go compiler recognizes the special case of map[string([]byte)] when used exactly
> +// in that pattern. See https://github.com/cosmos/cosmos-sdk/issues/8693.
> +var addMu sync.Mutex
> +var addrStrMemoize = make(map[string]string)
>
> Should there be some sort of eviction of the memoized addresses
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/cosmos/cosmos-sdk/pull/8694#pullrequestreview-598346574>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABFL3V2IWP4Y4PZPBLR43QLTAYIOXANCNFSM4YFPGATA>
> .
>