@robert-zaremba 2021-02-25
I would like to second on @marbar3778  comment - introducing a global cache with mutex may lead to memory problems. 
I don't think this should be committed without eviction mechanism. Here, the memory can growth substantially. 
Thoughts?
cc: @amaurym , @aaronc 

Also, for small optimization - we could move mutex is only when we set a value. (here the entry is updated at most once).