@odeke-em 2021-02-25
Robert, I think a fully fledged mechanism without progressively
benchmarking is a premature approach that’ll burn time and resources. We
start by eliminating big problems, keep profiling away the top issues, and
with live traffic keep isolating. I mentioned that am progressively working
on more changes but this suffices for now.

On Thu, Feb 25, 2021 at 3:32 AM Robert Zaremba <notifications@github.com>
wrote:

> I would like to second on @marbar3778 <https://github.com/marbar3778>
> comment - introducing a global cache with mutex may lead to memory problems.
> I don't think this should be committed without eviction mechanism. Here,
> the memory can growth substantially.
> Thoughts?
> cc: @amaurym <https://github.com/amaurym> , @aaronc
> <https://github.com/aaronc>
>
> Also, for small optimization - we could move mutex is only when we set a
> value. (here the entry is updated at most once).
>
> —
> You are receiving this because you modified the open/close state.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/cosmos/cosmos-sdk/pull/8694#issuecomment-785828838>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABFL3V5TVV4GVQY4YGLZZKLTAYYLZANCNFSM4YFPGATA>
> .
>