@robert-zaremba 2021-02-25
Totally agree for iterative approach. However I'm not convinced that this is change - we are doing a microbenchmark without looking at the bigger impact this can cause (not menagable memory growth).