@robert-zaremba 2021-02-25
To be more specific: IMHO micro-optimization of `0.02 ms` without more data (both how it impacts the overall performance and storage growth) is not worth taking a risk of potential memory leak.