@amaurym 2021-02-25
I agree with @robert-zaremba here. Before making this change, I'd like to see some profiling on how this map grows on e.g. Gaia for a couple of days. If it grows in an unexpected manner, I prefer to revert this change, have slightly sub-optimized code, but not introducing memory leaks on live chains.