@odeke-em 2021-02-25
Am definitely fine with a revert but is incorrect to call this a “micro
optimization” as clearly this is profile guided changes to pop off
something that was consuming RAM, this isn’t about CPU time, it clearly is
RAM taking down from >2GiB down to a lower value that disappear, and that’s
what profile guided development entails. So again this isn’t a CPU
micro-optimization but a RAM consumption problem as I showed in the issue.

On Thu, Feb 25, 2021 at 10:27 AM Amaury <notifications@github.com> wrote:

> I agree with @robert-zaremba <https://github.com/robert-zaremba> here.
> Before making this change, I'd like to see some profiling on how this map
> grows on e.g. Gaia for a couple of days. If it grows in an unexpected
> manner, I prefer to revert this change, have slightly sub-optimized code,
> but not introducing memory leaks on live chains.
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/cosmos/cosmos-sdk/pull/8694#issuecomment-786109132>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABFL3V5XCXK7UHTR2WCYJ4LTA2JBTANCNFSM4YFPGATA>
> .
>