@robert-zaremba 2021-02-26
> it clearly is RAM taking down from >2GiB down to a lower value that disappear

RAM optimization is definitely interesting. How did you benchmark this? In #8717  I changed map to a lru structure and would be interesting to repeat some flow to find a good _cache size_ parameter.